Folder id input, collision GUI, and more
Now attempts to match the input folder name to a folder ID. If this
fails, will go to name comparison. In the event of multiple folders of
the same name, opens a window asking the user to select one. Checking
for the default folder now occurs only when necessary instead of on
every execution. Increased verbosity of scheduling status, adding an
indication of whether or not a recording was overwritten.